use driver db in language server
