+++ crates/language-server/src/functionality/mod.rs
@@ -1,5 +1,4 @@
 mod capabilities;
-pub(super) mod diagnostics;
 pub(super) mod goto;
 pub(super) mod handlers;
 pub(super) mod hover;
